reverted variant application
